Move runtime tests to corresponding sub-modules.
